--- mysten-mindsV2/locale/en.json
@@ -1,72 +1,0 @@
-{
-  "connect_button": "Connect Wallet",
-  "hero_title": "Your AI Guide to the SUI Ecosystem",
-  "hero_description": "Get instant answers to all your questions about SUI blockchain, Move programming, and the entire Mysten Labs ecosystem.",
-  "ask_question": "Ask a Question",
-  "robot_interface_title": "MysteinMinds AI Interface",
-  "robot_idle_message": "Hi! Ask me anything about SUI blockchain.",
-  "robot_thinking_message": "Hmm, let me think about that",
-  "robot_speaking_message": "Great question! I'm processing your request...",
-  "robot_scanning_message": "Scanning for wallet connections...",
-  "ask_placeholder": "Ask anything about SUI blockchain...",
-  "suggested_questions": "Suggested Questions",
-  "question_what_is_sui": "What is SUI blockchain?",
-  "question_create_wallet": "How do I create a SUI wallet?",
-  "question_move_programming": "Explain Move programming language",
-  "question_sui_nfts": "What are SUI NFTs?",
-  "status_idle": "READY",
-  "status_thinking": "PROCESSING",
-  "status_speaking": "RESPONDING",
-  "status_scanning": "SCANNING",
-  "connect_wallet_title": "Connect Your Wallet",
-  "connect_wallet_description": "Select a wallet to connect to the MysteinMinds platform",
-  "wallet_connected_title": "Wallet Connected",
-  "wallet_connected_description": "Your wallet is connected to the MysteinMinds platform",
-  "wallet_address": "Address",
-  "wallet_balance": "Balance",
-  "view_on_explorer": "View on Explorer",
-  "disconnect_wallet": "Disconnect Wallet",
-
-  "swipe_up_hint": "Swipe up to ask a question",
-  "swipe_down_hint": "Swipe down to return",
-  "pull_to_refresh": "Pull down to refresh",
-  "release_to_refresh": "Release to refresh",
-  "refreshing": "Refreshing...",
-
-  "wallet_title": "Wallet",
-  "total_balance": "Total Balance",
-  "send": "Send",
-  "swap": "Swap",
-  "receive": "Receive",
-  "recent_transactions": "Recent Transactions",
-  "view_all_transactions": "View All Transactions",
-
-  "install_app": "Install App",
-  "add_to_home_screen": "Add to Home Screen",
-
-  "connect_wallet_to_chat": "Connect wallet to access chat",
-  "connect_wallet_placeholder": "Connect wallet to ask questions...",
-  "auth_required_title": "Authentication Required",
-  "auth_required_description": "You need to connect your wallet to use the chat",
-  "auth_required_message": "To access the chat and save your conversation history, please connect your wallet first.",
-  "connect_wallet": "Connect Wallet",
-
-  "chat_history": "Chat History",
-  "no_chats_yet": "No Chats Yet",
-  "start_new_chat_prompt": "Start a new chat to begin asking questions",
-  "new_chat": "New Chat",
-  "no_messages": "No messages yet",
-  "delete_chat": "Delete Chat",
-  "delete_chat_confirm": "Are you sure you want to delete this chat?",
-  "delete": "Delete",
-  "cancel": "Cancel",
-
-  "community": "Community",
-  "telegram_communities": "Telegram Communities",
-  "join_group": "Join Group",
-  "loading_telegram_groups": "Loading Telegram groups...",
-  "try_again": "Try Again",
-  "no_groups_found": "No groups found",
-  "no_groups_in_category": "There are no groups in this category",
-  "suggested_telegram_groups": "Suggested Telegram Groups"
-}
