--- mysten-mindsV2/locale/fr.json
@@ -1,29 +1,0 @@
-{
-  "connect_button": "Connecter Portefeuille",
-  "hero_title": "Votre Guide IA pour l'Écosystème SUI",
-  "hero_description": "Obtenez des réponses instantanées à toutes vos questions sur la blockchain SUI, la programmation Move et l'ensemble de l'écosystème Mysten Labs.",
-  "ask_question": "Poser une Question",
-  "robot_interface_title": "Interface IA MysteinMinds",
-  "robot_idle_message": "Bonjour! Posez-moi n'importe quelle question sur la blockchain SUI.",
-  "robot_thinking_message": "Hmm, laissez-moi réfléchir",
-  "robot_speaking_message": "Excellente question! Je traite votre demande...",
-  "robot_scanning_message": "Recherche de connexions de portefeuille...",
-  "ask_placeholder": "Demandez n'importe quoi sur la blockchain SUI...",
-  "suggested_questions": "Questions Suggérées",
-  "question_what_is_sui": "Qu'est-ce que la blockchain SUI?",
-  "question_create_wallet": "Comment créer un portefeuille SUI?",
-  "question_move_programming": "Expliquez le langage de programmation Move",
-  "question_sui_nfts": "Que sont les NFT SUI?",
-  "status_idle": "PRÊT",
-  "status_thinking": "TRAITEMENT",
-  "status_speaking": "RÉPONSE",
-  "status_scanning": "ANALYSE",
-  "connect_wallet_title": "Connecter Votre Portefeuille",
-  "connect_wallet_description": "Sélectionnez un portefeuille pour vous connecter à la plateforme MysteinMinds",
-  "wallet_connected_title": "Portefeuille Connecté",
-  "wallet_connected_description": "Votre portefeuille est connecté à la plateforme MysteinMinds",
-  "wallet_address": "Adresse",
-  "wallet_balance": "Solde",
-  "view_on_explorer": "Voir sur l'Explorateur",
-  "disconnect_wallet": "Déconnecter le Portefeuille"
-}
