--- mysten-mindsV2/locale/sw.json
@@ -1,29 +1,0 @@
-{
-  "connect_button": "Unganisha Wallet",
-  "hero_title": "Mwongozo Wako wa AI kwa Mfumo wa SUI",
-  "hero_description": "Pata majibu ya haraka kwa maswali yako yote kuhusu blockchain ya SUI, programu ya Move, na mfumo mzima wa Mysten Labs.",
-  "ask_question": "Uliza Swali",
-  "robot_interface_title": "Kiolesura cha AI cha MysteinMinds",
-  "robot_idle_message": "Habari! Niulize chochote kuhusu blockchain ya SUI.",
-  "robot_thinking_message": "Hmm, acha nifikirie",
-  "robot_speaking_message": "Swali zuri! Nasindika ombi lako...",
-  "robot_scanning_message": "Inatafuta unganisho wa wallet...",
-  "ask_placeholder": "Uliza chochote kuhusu blockchain ya SUI...",
-  "suggested_questions": "Maswali Yaliyopendekezwa",
-  "question_what_is_sui": "Blockchain ya SUI ni nini?",
-  "question_create_wallet": "Ninawezaje kuunda pochi ya SUI?",
-  "question_move_programming": "Eleza lugha ya programu ya Move",
-  "question_sui_nfts": "NFT za SUI ni nini?",
-  "status_idle": "TAYARI",
-  "status_thinking": "INASINDIKA",
-  "status_speaking": "INAJIBU",
-  "status_scanning": "INASKANISHA",
-  "connect_wallet_title": "Unganisha Wallet Yako",
-  "connect_wallet_description": "Chagua wallet ya kuunganisha kwenye jukwaa la MysteinMinds",
-  "wallet_connected_title": "Wallet Imeunganishwa",
-  "wallet_connected_description": "Wallet yako imeunganishwa kwenye jukwaa la MysteinMinds",
-  "wallet_address": "Anwani",
-  "wallet_balance": "Salio",
-  "view_on_explorer": "Tazama kwenye Explorer",
-  "disconnect_wallet": "Tenganisha Wallet"
-}
