+++ mysten-mindsV2/package.json
@@ -62,6 +62,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.54.1",
     "react-i18next": "latest",
+    "react-markdown": "^10.1.0",
     "react-resizable-panels": "^2.1.7",
     "react-swipeable": "latest",
     "recharts": "2.15.0",
